# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Memtrace API
  description: |
    LLM-agnostic memory layer for AI agents. Store, recall, and search
    temporal memories. Works with ChatGPT, Claude, Gemini, DeepSeek, Llama — any LLM.
  version: 0.1.0
  contact:
    name: Basekick Labs
    url: https://github.com/Basekick-Labs/memtrace
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9100
    description: Local development
...
paths:
...
  /api/v1/sessions:
    post:
      operationId: createSession
      tags: [Sessions]
      summary: Create a new session
      security:
        - ApiKeyHeader: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: listSessions
      tags: [Sessions]
      summary: List sessions
      security:
        - ApiKeyHeader: []
        - BearerAuth: []
      parameters:
        - name: agent_id
          in: query
          description: Filter by agent ID
          schema:
            type: string
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Session"
                  count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{id}:
    get:
      operationId: getSession
      tags: [Sessions]
      summary: Get a session by ID
      security:
        - ApiKeyHeader: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Session found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      operationId: updateSession
      tags: [Sessions]
      summary: Update session status or metadata
      description: |
        Setting status to `closed` will automatically populate `closed_at`.
        Metadata is merged with existing values.
      security:
        - ApiKeyHeader: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionRequest"
      responses:
        "200":
          description: Session updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{id}/memories:
    get:
      operationId: listSessionMemories
      tags: [Sessions]
      summary: List memories for a session
      security:
        - ApiKeyHeader: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: memory_type
          in: query
          schema:
            type: string
            enum: [episodic, session, decision, entity]
        - name: since
          in: query
          description: "Relative (e.g. 2h, 7d) or ISO 8601"
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{id}/context:
    post:
      operationId: getSessionContext
      tags: [Sessions]
      summary: Get LLM-ready session context
      description: |
        Returns memories for a session formatted as markdown, grouped by type
        (actions, decisions, errors, other). Ready to inject into any LLM prompt.
      security:
        - ApiKeyHeader: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextOptions"
      responses:
        "200":
          description: Session context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionContext"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

# ──────────────────────────────────────────────
# Components
# ──────────────────────────────────────────────
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: "Memtrace API key (mtk_ prefix)"
    BearerAuth:
      type: http
      scheme: bearer
      description: "Bearer token using mtk_ API key"
  parameters:
...
    SessionId:
      name: id
      in: path
      required: true
      description: Session ID
      schema:
        type: string
        example: sess_xyz789
    Limit:
      name: limit
      in: query
      description: Maximum number of results
      schema:
        type: integer
        default: 100
        minimum: 1
        maximum: 1000
...
    Order:
      name: order
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Memory schemas ────────────────────────
    Memory:
      type: object
      required: [time, org_id, agent_id, memory_type, event_type, content]
      properties:
        time:
          type: string
          format: date-time
          example: "2026-02-08T10:30:45Z"
        org_id:
          type: string
          example: org_default
        agent_id:
          type: string
          example: agent_web_crawler
        session_id:
          type: string
          example: sess_abc123
        memory_type:
          type: string
          enum: [episodic, session, decision, entity]
          example: episodic
        event_type:
          type: string
          example: page_crawled
        content:
          type: string
          example: "Crawled https://example.com — found 3 product pages"
        metadata:
          type: object
          additionalProperties: true
          example: { "url": "https://example.com", "pages_found": 3 }
        tags:
          type: array
          items:
            type: string
          example: ["crawling", "products"]
        dedup_key:
          type: string
        importance:
          type: number
          format: double
          minimum: 0
          maximum: 1
          example: 0.7
        parent_id:
          type: string
...
    MemoryList:
      type: object
      properties:
        memories:
          type: array
          items:
            $ref: "#/components/schemas/Memory"
        count:
          type: integer
        has_more:
          type: boolean

    # ── Search schemas ────────────────────────
...
    Session:
      type: object
      required: [id, org_id, agent_id, status, created_at, updated_at]
      properties:
        id:
          type: string
          example: sess_xyz789
        org_id:
          type: string
          example: org_default
        agent_id:
          type: string
          example: agent_abc123
        status:
          type: string
          enum: [active, closed]
          example: active
        metadata:
          type: object
          additionalProperties: true
          example: { "task": "product_research" }
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
          nullable: true
    CreateSessionRequest:
      type: object
      required: [agent_id]
      properties:
        agent_id:
          type: string
          example: agent_abc123
        metadata:
          type: object
          additionalProperties: true
          example: { "task": "product_research" }
    UpdateSessionRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, closed]
        metadata:
          type: object
          additionalProperties: true
    ContextOptions:
      type: object
      properties:
        max_tokens:
          type: integer
          description: Approximate max token budget for context
        include_types:
          type: array
          items:
            type: string
            enum: [episodic, session, decision, entity]
        since:
          type: string
          description: "Relative (e.g. 4h) or ISO 8601. Default: 24h"
    SessionContext:
      type: object
      properties:
        session_id:
          type: string
          example: sess_xyz789
        context:
          type: string
          description: Markdown-formatted LLM-ready context
          example: "## Session Context (sess_xyz789)\n\n### Recent Actions (3)\n- [2026-02-08T14:22:10Z] page_crawled: Crawled https://example.com...\n"
        memory_count:
          type: integer
          example: 3